[HUBS] Use visible instead of layers for intersectObject
Older versions of the ThreeJS Raycast API ignored invisible objects.
This was changed to use layers instead, but Hubs currently still relies
on the old behavior. Changing this would require a bunch of code changes so
we retain this old behavior for now.

Behavior changed in this commit:
https://github.com/mrdoob/three.js/commit/4ce5ba209c6d1712e3728726b0126c34c237a626#diff-e9ba2746300f68606637fd8769a722f23924e93ecb2aefead0163ba251d3e888

More context:
https://github.com/mrdoob/three.js/issues/14700