[HUBS] Do not version control built assets
We want to maintain a clean set of patch commits for our threejs fork.
Getting rid of the built assets makes it easier to apply these cleanly.

Instead we build the only asset we care about three.module.js in the npm
"prepare" hook, which will run when the package is installed. Also adds
a "watch-esm" script that will watch for changes and rebuild only
three.module.js, which can bee useful for local development.

This commit may have a conflict when we rebase, but we can simply throw
away the remote changes here once. And then the rest of the commits
should apply cleanly, rather than having to resolve issues on the built
assets for each commit.